langweave::language_detector_trait
Trait LanguageDetectorTrait

pub trait LanguageDetectorTrait: Send + Sync {
    // Required methods
    fn detect(&self, text: &str) -> Result<String, I18nError>;
    fn detect_async<'life0, 'life1, 'async_trait>(
        &'life0 self,
        text: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<String, I18nError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

A trait for implementing custom language detection methods.

Required Methods

fn detect(&self, text: &str) -> Result<String, I18nError>

Detects the language of the given text synchronously.

Arguments

text - A string slice that holds the text to analyze.

Returns

Result<String, I18nError> - The detected language code if successful, or an error if detection fails.

fn detect_async<'life0, 'life1, 'async_trait>( &'life0 self, text: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<String, I18nError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Detects the language of the given text asynchronously.

Arguments

text - A string slice that holds the text to analyze.

Returns

Result<String, I18nError> - The detected language code if successful, or an error if detection fails.